+++ src/components/BreedCard/index.ts
@@ -1,0 +1,1 @@
+export { default as BreedCard } from './BreedCard';
