+++ src/components/BreedDetails/index.ts
@@ -1,0 +1,1 @@
+export { default as BreedDetails } from './BreedDetails';
