+++ src/components/BreedList/index.ts
@@ -1,0 +1,1 @@
+export { default as BreedList } from './BreedList';
