+++ src/components/Flyout/index.ts
@@ -1,0 +1,1 @@
+export { default as Flyout } from './Flyout';
