+++ src/components/NoResultsPlaceholder/index.ts
@@ -1,0 +1,1 @@
+export { default as NoResultsPlaceholder } from './NoResultsPlaceholder';
