+++ src/components/Pagination/index.ts
@@ -1,0 +1,1 @@
+export { default as Pagination } from './Pagination';
