+++ src/components/PopUpMessage/index.ts
@@ -1,0 +1,1 @@
+export { default as PopUpMessage } from './PopUpMessage';
