+++ src/components/SearchForm/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchForm } from './SearchForm';
