+++ src/components/index.ts
@@ -1,0 +1,11 @@
+export * from './BreedCard';
+export * from './BreedDetails';
+export * from './BreedList';
+export * from './ErrorBoundary';
+export * from './Flyout';
+export * from './Header';
+export * from './Loader';
+export * from './NoResultsPlaceholder';
+export * from './Pagination';
+export * from './PopUpMessage';
+export * from './SearchForm';
