+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
+import { ErrorBoundary } from '@/components/';
 import { ThemeProvider } from '@/contexts';
 import App from './App.tsx';
