+++ src/pages/PageNotFound/index.ts
@@ -1,0 +1,1 @@
+export { default as PageNotFound } from './PageNotFound';
