+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './Home';
+export * from './About';
+export * from './PageNotFound';
